Transform a goal model to corresponding CSP# code
Goal models are tree-based model that are used to represents the
refinement of a software system requirements. To automatically generate
systems'  behavior, we propose to use PAT to simulate the business logic
implied by the goal model. This project primarly aims to transform goal
model to the corresponding CSP# that can be compiled in PAT.